+++ vite.config.ts
@@ -35,7 +35,7 @@ export default defineConfig(({ mode }) => {
         ],
       }),
       VitePWA({
-        registerType: "autoUpdate",
+        registerType: "prompt",
         strategies: "injectManifest",
         srcDir: "src",
         filename: "sw.ts",